Cleaning up utilities files.
